==> 1. Databehandling.ipynb <==
# Importing the required library
import pandas as pd

# Reading the Excel file into a Pandas DataFrame
df = pd.read_excel('Data/Data_BA.xlsx')

# Displaying the first 5 rows of the DataFrame to verify that the data has been read correctly
df


# Removing rows where 'PrisIAlt' is NaN (Not a Number) and 'status' is 'åben'
df.dropna(subset=['PrisIAlt'], inplace=True)
df = df[df['status'] != 'åben']

# Displaying the first few rows of the modified DataFrame to verify the changes
df

# One-hot encode the categorical control variables without dropping the first dummy variable
encoded_data = pd.get_dummies(df, columns=['Indsats', 'Underretter', 'BaggrundBarnet', 'BaggrundHjemmet', 'Område'])

# Check the shape of the cleaned and encoded data
print(encoded_data.shape)



# 1. Identify the columns that contain the ¤ symbol
cols_with_symbol = [col for col in encoded_data.columns if '¤' in col]

# 2. For each identified column, split its values based on the ¤ symbol and create new columns
for col in cols_with_symbol:
    # Split the column name by the ¤ symbol
    split_names = col.split('¤')
    
    # For each split name, create a new column if it doesn't exist, and copy the values
    for name in split_names:
        new_col_name = col.split('_')[0] + '_' + name
        if new_col_name not in encoded_data.columns:
            encoded_data[new_col_name] = 0
        
        # Add the values from the original column to the new column
        encoded_data.loc[encoded_data[col] == 1, new_col_name] = 1

# 3. Drop the original columns that contained the ¤ symbol
encoded_data.drop(columns=cols_with_symbol, inplace=True)

# Check the resulting dataframe
print(encoded_data.columns)


try:
    encoded_data['Underretter_Anden_Kommune'] = encoded_data['Underretter_Anden kommunal afdeling'] + encoded_data['Underretter_Modtaget fra anden kommune']
    print("Column 'Underretter_Anden_Kommune' created.")
    
    unique_vals = encoded_data['Underretter_Anden_Kommune'].unique()
    print(f"Unique values in 'Underretter_Anden_Kommune': {unique_vals}")
    
    encoded_data.drop(['Underretter_Anden kommunal afdeling', 'Underretter_Modtaget fra anden kommune'], axis=1, inplace=True)
    print("Old columns dropped.")
    
except Exception as e:
    print(f"An error occurred: {e}")


# 2nd Combination
try:
    encoded_data['Underretter_Sundhedsvæsen'] = encoded_data[['Underretter_Øvrige sundhedsvæsen', 'Underretter_Læge', 'Underretter_Sundhedsplejerske', 'Underretter_Sygehus']].sum(axis=1)
    print("Column 'Underretter_Sundhedsvæsen' created.")
    encoded_data.drop(['Underretter_Øvrige sundhedsvæsen', 'Underretter_Læge', 'Underretter_Sundhedsplejerske', 'Underretter_Sygehus'], axis=1, inplace=True)
    print("Old columns for 'Underretter_Sundhedsvæsen' dropped.")
except Exception as e:
    print(f"An error occurred with 'Underretter_Sundhedsvæsen': {e}")

# 3rd Combination
try:
    encoded_data['Underretter_Barnets netværk'] += encoded_data['Underretter_Vurdering af underretning § 155']
    print("Column 'Underretter_Barnets netværk' updated.")
    encoded_data.drop(['Underretter_Vurdering af underretning § 155'], axis=1, inplace=True)
    print("Old column for 'Underretter_Barnets netværk' dropped.")
except Exception as e:
    print(f"An error occurred with 'Underretter_Barnets netværk': {e}")

# 4th Combination
try:
    encoded_data['Underretter_Krisecenter'] += encoded_data['Underretter_Døgnvagt']
    print("Column 'Underretter_Krisecenter' updated.")
    encoded_data.drop(['Underretter_Døgnvagt'], axis=1, inplace=True)
    print("Old column for 'Underretter_Krisecenter' dropped.")
except Exception as e:
    print(f"An error occurred with 'Underretter_Krisecenter': {e}")

# 5th Combination
try:
    encoded_data['Underretter_Daginstitution/Dagpleje'] += encoded_data['Underretter_Daginstitution']
    print("Column 'Underretter_Daginstitution/Dagpleje' updated.")
    encoded_data.drop(['Underretter_Daginstitution'], axis=1, inplace=True)
    print("Old column for 'Underretter_Daginstitution/Dagpleje' dropped.")
except Exception as e:
    print(f"An error occurred with 'Underretter_Daginstitution/Dagpleje': {e}")

# 6th Combination
try:
    encoded_data['Underretter_Socialrådgiver'] = encoded_data['Underretter_Kontaktperson'] + encoded_data['Underretter_Familiebehandling']
    print("Column 'Underretter_Socialrådgiver' created.")
    encoded_data.drop(['Underretter_Kontaktperson', 'Underretter_Familiebehandling'], axis=1, inplace=True)
    print("Old columns for 'Underretter_Socialrådgiver' dropped.")
except Exception as e:
    print(f"An error occurred with 'Underretter_Socialrådgiver': {e}")


encoded_data.drop(columns=['Indsatsparagraf', 'Leverandør', 'Haendelsesdato', 'sag_start', 'sag_slut'], inplace=True)


num_columns = encoded_data.shape[1]
print(f"The dataframe has {num_columns} columns.")

# Save the modified DataFrame to a new Excel file
encoded_data.to_excel('Data/Data_BA_modified.xlsx', index=False)
